requirements.txt
requests
beautifulsoup4
readability-lxml
sentence-transformers
faiss-cpu
fastapi
uvicorn
jsonlines
ollama
streamlit